+++ autodiff/forward/dual.hpp
@@ -31,4 +31,4 @@
 
 // autodiff includes
 #include <autodiff/forward/dual/dual.hpp>
-#include <autodiff/forward/common/derivative.hpp>
+#include <autodiff/forward/utils/derivative.hpp>
