+++ autodiff/forward/dual/dual.hpp
@@ -37,8 +37,8 @@
 #include <utility>
 
 // autodiff includes
-#include <autodiff/utils/aliases.hpp>
-#include <autodiff/utils/traits.hpp>
+#include <autodiff/common/numbertraits.hpp>
+#include <autodiff/common/meta.hpp>
 
 namespace autodiff {
 namespace detail {
