+++ autodiff/forward/real/eigen.hpp
@@ -34,8 +34,8 @@
 
 // autodiff includes
 #include <autodiff/forward/real.hpp>
-#include <autodiff/forward/common/gradient.hpp>
-#include <autodiff/utils/eigenmacros.hpp>
+#include <autodiff/forward/utils/gradient.hpp>
+#include <autodiff/common/eigen.hpp>
 
 //------------------------------------------------------------------------------
 // SUPPORT FOR EIGEN MATRICES AND VECTORS OF REAL
