+++ tests/common/meta.test.cpp
@@ -2,7 +2,7 @@
 #include <catch2/catch.hpp>
 
 // autodiff includes
-#include <autodiff/utils/meta.hpp>
+#include <autodiff/common/meta.hpp>
 using namespace autodiff;
 
 TEST_CASE("testing autodiff meta utilities", "[utils][meta]")
